refactor: 인기 아고라 쿼리 및 swiper 로직 분리
